:zap: Refactor project structure and add field for notification construction
