+++ nodes/Ntfy/data/emojis.json
