+++ nodes/Ntfy/fields/generalFields.ts
@@ -1,40 +1,16 @@
 import { INodeProperties } from 'n8n-workflow';
 
-export const mainFields: INodeProperties[] = [
-	{
-		displayName: 'Use Custom Server',
-		name: 'useCustomServer',
-		description:
-			'Whether to use a custom server. If not, the default server (ntfy.sh) will be used.',
-		type: 'boolean',
-		default: false,
-	},
-	{
-		displayName: 'Server URL',
-		name: 'serverUrl',
-		description: 'The URL of the ntfy.sh server',
-		type: 'string',
-		default: '',
-		placeholder: 'https://ntfy.sh',
-		displayOptions: {
-			show: {
-				useCustomServer: [true],
-			},
-		},
-	},
-	{
-		displayName: 'Topic',
-		name: 'topic',
-		type: 'string',
-		default: '',
-		placeholder: 'mytopic',
-		required: true,
-	},
+export const generalFields: INodeProperties[] = [
 	{
 		displayName: 'Title',
 		name: 'title',
 		type: 'string',
 		default: '',
+		displayOptions: {
+			show: {
+				constructNotification: ['generalFields'],
+			},
+		},
 	},
 	{
 		displayName: 'Priority',
@@ -64,6 +40,11 @@ export const mainFields: INodeProperties[] = [
 			},
 		],
 		noDataExpression: true,
+		displayOptions: {
+			show: {
+				constructNotification: ['generalFields'],
+			},
+		},
 	},
 	{
 		displayName: 'Emojis/Tags',
@@ -106,6 +87,11 @@ export const mainFields: INodeProperties[] = [
 				],
 			},
 		],
+		displayOptions: {
+			show: {
+				constructNotification: ['generalFields'],
+			},
+		},
 	},
 	{
 		displayName: 'Message',
@@ -114,6 +100,11 @@ export const mainFields: INodeProperties[] = [
 		default: '',
 		typeOptions: {
 			rows: 7,
+		},
+		displayOptions: {
+			show: {
+				constructNotification: ['generalFields'],
+			},
 		},
 	},
 ];
